+++ examples/Demo/FluentUI.Demo.Client/Infrastructure/ServiceCollectionExtensions.cs
@@ -2,7 +2,6 @@
 // MIT License - Copyright (c) Microsoft Corporation. All rights reserved.
 // ------------------------------------------------------------------------
 
-using FluentUI.Demo.Client.Infrastructure.ConsoleLog;
 using FluentUI.Demo.DocViewer;
 
 namespace FluentUI.Demo.Client;
@@ -57,9 +56,6 @@ public static class ServiceCollectionExtensions
     {
         // Add HttpClient for the "CatchAll.razor" or "DocViewer.razor" page
         services.AddHttpClient();
-
-        // Add Console Log Service
-        services.AddScoped<ConsoleLogService>();
 
         // Documentation services
         services.AddDocViewer(options =>
