+++ examples/Demo/FluentUI.Demo.Client/Layout/DemoMainLayout.razor
@@ -1,4 +1,4 @@
-﻿@using FluentUI.Demo.Client.Infrastructure.ConsoleLog
+﻿@using FluentUI.Demo.DocViewer.Components
 @inject NavigationManager Navigation
 <PageTitle>@App.PageTitle("Demo")</PageTitle>
