+++ examples/Tools/FluentUI.Demo.DocViewer/Components/ConsoleLogProvider.razor.css
