+++ examples/Tools/FluentUI.Demo.DocViewer/Components/MarkdownViewer.razor.cs
@@ -43,6 +43,8 @@ public partial class MarkdownViewer
     /// <summary />
     protected override async Task OnInitializedAsync()
     {
+        Console.WriteLine("OnInitializedAsync");
+
         // Markdown
         var page = DocViewerService.FromRoute(Route);
